+++ hardhat/package.json
@@ -27,6 +27,7 @@
     "hardhat": "^2.10.1"
   },
   "dependencies": {
+    "@chainlink/contracts": "^0.4.2",
     "dotenv": "^16.0.1"
   }
 }
